requirements.txt
Django>=1.8,<2.0
psycopg2
djangorestframework
requests
Pillow
